# cmd/payment/docs/swagger.yaml
basePath: /
definitions:
  model.CreateTransactionRequest:
    properties:
      amount:
        type: number
      currency:
        type: string
      customer_id:
        type: integer
      hash_credit_card:
        type: string
      merchant_id:
        type: integer
    type: object
...
  model.Transaction:
    properties:
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      customer_id:
        type: integer
      hash_credit_card:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      state:
        type: string
      token:
        type: string
      updated_at:
        type: string
    type: object
  model.TransactionUpdateDTO:
    properties:
      amount:
        description: omitempty permite que el campo sea opcional
        type: number
      currency:
        type: string
      state:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para el manejo de pagos
  title: API de Pagos
  version: "1"
paths:
...
  /api/transaction:
    post:
      consumes:
      - application/json
      description: Crea una nueva transacción con los datos proporcionados
      parameters:
      - description: Datos de la Transacción
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Transaction'
      security:
      - ApiKeyAuth: []
      summary: Crea una nueva transacción
      tags:
      - transaction
  /api/transaction/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene una transaccion por su ID
      parameters:
      - description: ID del Transaction
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaccion encontrada
          schema:
            $ref: '#/definitions/model.Transaction'
        "404":
          description: Transaccion no encontrada
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene una transacción
      tags:
      - transaction
    put:
      consumes:
      - application/json
      description: Actualiza una transacción existente; solo se permite si el estado
        es "pending"
      parameters:
      - description: ID de la Transacción
        in: path
        name: id
        required: true
        type: integer
      - description: Datos de la Transacción a actualizar
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.TransactionUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Transacción actualizada
        "400":
          description: Entrada inválida
        "404":
          description: Transacción no encontrada
      security:
      - ApiKeyAuth: []
      summary: Actualiza una transacción
      tags:
      - transaction
  /api/transaction/process:
    post:
      consumes:
      - application/json
      description: Procesa una transacción dada, capturándola o reembolsándola basado
        en el tipo de transacción especificado.
      parameters:
      - description: Token de la transacción
        in: query
        name: token
        required: true
        type: string
      - description: Tipo de transacción (capture, refund)
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transacción procesada
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Error de solicitud - Transacción ya procesada, tipo de transacción
            inválido, tarjeta de crédito inválida, etc.
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transacción no encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Procesa una transacción
      tags:
      - transaction
  /api/transaction/token/{token}:
    get:
      consumes:
      - application/json
      description: Obtiene una transacción por su token único
      parameters:
      - description: Token de la Transacción
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Transaction'
        "404":
          description: Transaction not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtiene una transacción por token
      tags:
      - transaction
...
swagger: "2.0"